+++ sereto/cli/utils.py
@@ -1,21 +1,12 @@
-import functools
-import os
-import pathlib
-import sys
-from collections.abc import Callable
 from enum import Enum
-from typing import Any, ParamSpec, TypeVar
+from typing import TypeVar
 
 import click
-import jinja2
-import pydantic
-from pydantic import validate_call
-from rich.markup import escape
+from rich.console import Console as RichConsole
 from rich.prompt import Prompt
 
 from sereto.cli.aliases import cli_aliases
-from sereto.cli.console import Console
-from sereto.exceptions import SeretoException
+from sereto.singleton import Singleton
 
 
 class AliasedGroup(click.Group):
@@ -61,111 +52,6 @@ class AliasedGroup(click.Group):
         return cmd.name, cmd, args
 
 
-# Param = ParamSpec("Param")
-# RetType = TypeVar("RetType")
-
-
-# def settings_require(vars: set[Literal["reports_path", "templates_path"]]
-#                      ) -> Callable[[Callable[Param, RetType]], Callable[Param, RetType]]:
-#     """Decorator marking the optional attributes of Settings as required.
-
-#     It load the CliContext through `click.get_current_context().obj`, which much already exist.
-
-#     Usage:
-#         `@settings_require(vars={"reports_path", "templates_path"})`
-#     """
-
-#     def decorator(func: Callable[Param, RetType]) -> Callable[Param, RetType]:
-#         @functools.wraps(func)
-#         def wrapper(*args: Param.args, **kwargs: Param.kwargs) -> RetType:
-#             if len(vars) == 0:
-#                 raise SeretoValueError("no parameter provided to settings decorator")
-#             cli_ctx: CliContext = click.get_current_context().obj
-
-#             for var_path in vars:
-#                 if getattr(cli_ctx.settings, var_path) is None:
-#                     Console().print(dedent(f"""\
-#                         [yellow]WARNING: {var_path!r} directory not defined
-
-#                         Your choices:
-#                         1) Fill in the path interactively through the active prompt.
-#                         The configuration regarding its location will be written
-#                         to "{get_settings_path()}".
-#                         2) Set environment variable SERETO_{var_path.upper()}
-#                         3) Manually create/edit the settings file.
-#                         You can use `sereto settings edit`.
-#                         """))
-#                     input: str = Prompt.ask(var_path, console=Console())
-#                     path = pathlib.Path(input).resolve()
-
-#                     if not path.exists():
-#                         if Confirm.ask(f"[yellow]Directory '{path}' does not exist. Create?",
-#                                        console=Console(), default=False):
-#                             path.mkdir(parents=True)
-#                         else:
-#                             raise SeretoPathError(f"{var_path!r} not provided")
-
-#                     setattr(cli_ctx.settings, var_path, path)
-#                     write_settings(cli_ctx.settings)
-
-#             return func(*args, **kwargs)
-#         return wrapper
-#     return decorator
-
-
-P = ParamSpec("P")
-R = TypeVar("R")
-
-
-def handle_exceptions(func: Callable[P, R]) -> Callable[P, R]:
-    """Decorator for pretty printing SeReTo exceptions in debug mode.
-
-    If the exception is a subclass of SeretoException and DEBUG environment variable is set to '1', the full exception
-    traceback will be printed with local variables shown.
-    """
-
-    @functools.wraps(func)
-    def outer_function(*args: P.args, **kwargs: P.kwargs) -> R:
-        try:
-            return func(*args, **kwargs)
-        except Exception as e:
-            if isinstance(e, SeretoException):
-                Console().print(f"[red]Error:[/red] {escape(str(e))}")
-            if os.environ.get("DEBUG", False):
-                Console().print_exception(show_locals=True, suppress=[click, jinja2, pydantic, pathlib])
-            else:
-                Console().print("\n[yellow]Set environment variable [blue]DEBUG=1[/blue] for more details.")
-            sys.exit(1)
-
-    return outer_function
-
-
-@validate_call
-def edit_list(list: list[Any], prompt: str) -> list[str]:
-    """Edit a list of objects using the user's default editor.
-
-    Args:
-        list: A list of objects (convertable to str) initially loaded to the editor. It remains unmodified.
-        prompt: A prompt to be displayed to the user before the editor is opened.
-
-    Returns:
-        New list entered by user to the text editor.
-
-    The edit_list function prompts the user to edit a list of strings using their default editor. It first prints the
-    given prompt to the console and waits for the user to press Enter. Then it opens the user's default editor with the
-    list of strings as the initial content. If the user saves and closes the editor, the edited content is returned as
-    a list of strings. If the user closes the editor without saving, an empty list is returned.
-
-    The function also removes any empty lines and lines starting with `#` from the edited content before returning it.
-    """
-    Console().print(prompt)
-    input("Press Enter to continue...")
-    res: str | None = click.edit("\n".join(list))
-    if res is None:
-        return []
-    return [line.strip() for line in res.split("\n") if len(line.strip()) > 0 and line[0] != "#"]
-
-
 EnumType = TypeVar("EnumType", bound=Enum)
 
 
@@ -173,3 +59,7 @@ def load_enum(enum: type[EnumType], prompt: str) -> EnumType:
     """Let user select a value from enum."""
     choice = Prompt.ask(prompt=prompt, choices=[e.value for e in enum])
     return enum(choice)
+
+
+class Console(RichConsole, metaclass=Singleton):
+    """Singleton wrapper around Rich's Console."""